Module Capnp_rpc_net.Auth

Vat-level authentication and encryption.

Unless your network provides a secure mechanism for establishing connections to other vats, where you can be sure of the identity of the other party, you'll probably want to enable cryptographic security.

Each vat (application instance) should generate a secret key when it is first deployed. For servers at least, this key must be saved to disk so that the server retains its identity over re-starts. Otherwise, clients will think it is an imposter and refuse to connect.

Clients that do not accept incoming connections, nor create SturdyRefs, can get away with creating a new key each time. However, it might be quicker to save and reload the key anyway.

type hash = [

| `SHA256

]

Supported hashes.

module Digest : sig ... end

module Secret_key : sig ... end