Module Capnp_rpc_net.Capnp_address

Handling of capnp:// URI format addresses. This code is shared between the unix and mirage networks.

module Location : sig ... end

include S.ADDRESS with type t = Location.t * Auth.Digest.t

type t = Location.t * Auth.Digest.t

A network address at which a vat can be reached.

val parse_uri : Uri.t -> (t * string, [> `Msg of string ]) Stdlib.result

parse_uri uri extracts from a URI the network address and service ID.

val to_uri : (t * string) -> Uri.t

to_uri (t, service_id) is a URI that can be parsed back into (t, service_id) by parse_uri.

val equal : t -> t -> bool

val digest : t -> Auth.Digest.t

How to verify that the correct address has been reached.

val pp : t Fmt.t