Module type S.VAT_NETWORK

Stretching capability references across a network link. Note: see Capnp_rpc_unix for a higher-level wrapper for this API.

type +'a capability

An 'a capability is a capability reference to a service of type 'a.

type flow

A bi-directional byte-stream.

type restorer

A function for restoring persistent capabilities from sturdy ref service IDs.

type +'a sturdy_ref

An off-line (persistent) capability.

type service_id

A (secret) token that identifies a persistent service within a vat and grants access to it.

module Network : NETWORK

module CapTP : sig ... end

Sharing capabilities over a network link.

module Vat : sig ... end

A Vat is an actor in the CapTP network, composed of objects that can call each other directly.