Module Capnp_rpc_net.Two_party_network

A network where the is only one other addressable party.

include S.NETWORK with type t = unit

module Types : Capnp_rpc_proto.S.NETWORK_TYPES

module Address : S.ADDRESS

type t = unit

val connect : t -> sw:Eio.Switch.t -> secret_key:Auth.Secret_key.t Stdlib.Lazy.t -> Address.t -> (Endpoint.t, [> `Msg of string ]) Stdlib.result

connect t ~sw ~secret_key address connects to address, proves ownership of secret_key (if TLS is being used), and returns the resulting endpoint. Returns an error if no connection can be established or the target fails to authenticate itself.

val parse_third_party_cap_id : Capnp_rpc.Private.Schema.Reader.pointer_t -> Types.third_party_cap_id