Module Tls_wrapper.Make

Parameters

Underlying : Mirage_flow.S

Signature

val connect_as_server : switch:Lwt_switch.t -> Underlying.flow -> Capnp_rpc_net.Auth.Secret_key.t option -> (Capnp_rpc_net.Endpoint.t, [> `Msg of string ]) Stdlib.result Lwt.t

val connect_as_client : switch:Lwt_switch.t -> Underlying.flow -> Capnp_rpc_net.Auth.Secret_key.t Stdlib.Lazy.t -> Capnp_rpc_net.Auth.Digest.t -> (Capnp_rpc_net.Endpoint.t, [> `Msg of string ]) Stdlib.result Lwt.t

connect_as_client ~switch underlying key digest is an endpoint using flow underlying. If digest requires TLS, it performs a TLS handshake. It uses key as its private key and checks that the server is the one required by auth.