Module Capnp_rpc_net__Two_party_network.Address

type t

A network address at which a vat can be reached.

val parse_uri : Uri.t -> (t * string, [> `Msg of string ]) Stdlib.result

parse_uri uri extracts from a URI the network address and service ID.

val to_uri : (t * string) -> Uri.t

to_uri (t, service_id) is a URI that can be parsed back into (t, service_id) by parse_uri.

val equal : t -> t -> bool

val digest : t -> Capnp_rpc_net.Auth.Digest.t

How to verify that the correct address has been reached.

val pp : t Fmt.t