Module Capnp_rpc_net

This package adds networking support, including TLS. It contains code common to capnp-rpc-unix and capnp-rpc-mirage. Libraries should not need to link against this package (just use capnp-rpc-lwt instead), since they generally shouldn't care whether services are local or remote.

module S : sig ... end

Module signatures.

module Endpoint : sig ... end

Send and receive capnp messages over a byte-stream.

module Two_party_network : sig ... end

A network where the is only one other addressable party.

module Auth : sig ... end

Vat-level authentication and encryption.

module Tls_wrapper : sig ... end

module Restorer : sig ... end

module type VAT_NETWORK = S.VAT_NETWORK with type 'a capability := 'a Capnp_rpc_lwt.Capability.t and type restorer := Restorer.t and type service_id := Restorer.Id.t and type 'a sturdy_ref := 'a Capnp_rpc_lwt.Sturdy_ref.t

Stretching capability references across a network link. Note: see Capnp_rpc_unix for a higher-level wrapper for this API.

module Networking (N : S.NETWORK) (Flow : Mirage_flow.S) : VAT_NETWORK with module Network = N and type flow = Flow.flow

Stretching capability references across a network link. Note: see Capnp_rpc_unix for a higher-level wrapper for this API.

module Capnp_address : sig ... end

Handling of capnp:// URI format addresses. This code is shared between the unix and mirage networks.